# openapi.yaml
openapi: 3.1.0
info:
  title: Code Workbench API
  version: 0.1.0
servers:
  - url: http://127.0.0.1:5174
paths:
...
  /api/repositories:
    post:
      operationId: createRepository
      tags: [repositories]
      requestBody: {$ref: '#/components/requestBodies/CreateRepository'}
      responses: {'201': {$ref: '#/components/responses/Repository'}, '400': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}
    get: {operationId: listRepositories, tags: [repositories], responses: {'200': {$ref: '#/components/responses/RepositoryList'}}}
  /api/sessions:
    post:
      operationId: createSession
      tags: [sessions]
      requestBody: {$ref: '#/components/requestBodies/CreateSession'}
      responses: {'201': {$ref: '#/components/responses/Session'}, '400': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}, '502': {$ref: '#/components/responses/Error'}}
    get: {operationId: listSessions, tags: [sessions], responses: {'200': {$ref: '#/components/responses/SessionList'}}}
  /api/sessions/{sessionId}:
    parameters: [{$ref: '#/components/parameters/SessionId'}]
    get: {operationId: getSession, tags: [sessions], responses: {'200': {$ref: '#/components/responses/Session'}, '404': {$ref: '#/components/responses/Error'}}}
  /api/sessions/{sessionId}/files:
    parameters: [{$ref: '#/components/parameters/SessionId'}, {$ref: '#/components/parameters/RelativePath'}]
    get: {operationId: getSessionFile, tags: [sessions], responses: {'200': {$ref: '#/components/responses/SourceFile'}, '400': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}}}
  /api/sessions/{sessionId}/intent:
    parameters: [{$ref: '#/components/parameters/SessionId'}]
    post:
      operationId: setSessionIntent
      tags: [sessions]
      requestBody: {$ref: '#/components/requestBodies/SessionIntent'}
      responses: {'200': {$ref: '#/components/responses/Session'}, '400': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}
  /api/sessions/{sessionId}/analysis-jobs:
    parameters: [{$ref: '#/components/parameters/SessionId'}]
    post: {operationId: queueAnalysisJob, tags: [jobs], requestBody: {$ref: '#/components/requestBodies/Provider'}, responses: {'200': {$ref: '#/components/responses/AgentJob'}, '202': {$ref: '#/components/responses/AgentJob'}, '400': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}, '502': {$ref: '#/components/responses/Error'}}}
  /api/candidates:
    get: {operationId: listCandidates, tags: [candidates], parameters: [{$ref: '#/components/parameters/SessionFilter'}, {$ref: '#/components/parameters/RepositoryFilter'}, {$ref: '#/components/parameters/StatusFilter'}, {$ref: '#/components/parameters/RiskFilter'}, {$ref: '#/components/parameters/ConfidenceFilter'}, {$ref: '#/components/parameters/CapabilityFilter'}], responses: {'200': {$ref: '#/components/responses/CandidateList'}}}
  /api/candidates/{candidateId}:
    parameters: [{$ref: '#/components/parameters/CandidateId'}]
    patch:
      operationId: updateCandidate
      tags: [candidates]
      requestBody: {$ref: '#/components/requestBodies/PatchCandidate'}
      responses: {'200': {$ref: '#/components/responses/Candidate'}, '400': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}
  /api/candidates/{candidateId}/approve:
    parameters: [{$ref: '#/components/parameters/CandidateId'}]
    post: {operationId: approveCandidate, tags: [candidates], requestBody: {$ref: '#/components/requestBodies/Reason'}, responses: {'200': {$ref: '#/components/responses/Candidate'}, '400': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}}
  /api/candidates/{candidateId}/reject:
    parameters: [{$ref: '#/components/parameters/CandidateId'}]
    post: {operationId: rejectCandidate, tags: [candidates], requestBody: {$ref: '#/components/requestBodies/Reason'}, responses: {'200': {$ref: '#/components/responses/Candidate'}, '400': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}}
  /api/candidates/{candidateId}/defer:
    parameters: [{$ref: '#/components/parameters/CandidateId'}]
    post: {operationId: deferCandidate, tags: [candidates], requestBody: {$ref: '#/components/requestBodies/Reason'}, responses: {'200': {$ref: '#/components/responses/Candidate'}, '400': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}}
  /api/candidates/{candidateId}/duplicate:
    parameters: [{$ref: '#/components/parameters/CandidateId'}]
    post: {operationId: duplicateCandidate, tags: [candidates], requestBody: {$ref: '#/components/requestBodies/Reason'}, responses: {'200': {$ref: '#/components/responses/Candidate'}, '400': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}}
  /api/candidates/{candidateId}/rescan:
    parameters: [{$ref: '#/components/parameters/CandidateId'}]
    post: {operationId: rescanCandidate, tags: [candidates], requestBody: {$ref: '#/components/requestBodies/OptionalReason'}, responses: {'200': {$ref: '#/components/responses/Candidate'}, '400': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}}
  /api/extraction-plans:
    post:
      operationId: createExtractionPlan
      tags: [extraction]
      requestBody: {$ref: '#/components/requestBodies/CreateExtractionPlan'}
      responses: {'201': {$ref: '#/components/responses/ExtractionPlan'}, '400': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}
  /api/extraction-plans/{planId}:
    parameters: [{$ref: '#/components/parameters/PlanId'}]
    get: {operationId: getExtractionPlan, tags: [extraction], responses: {'200': {$ref: '#/components/responses/ExtractionPlan'}, '404': {$ref: '#/components/responses/Error'}}}
  /api/extraction-plans/{planId}/jobs:
    parameters: [{$ref: '#/components/parameters/PlanId'}]
    post: {operationId: queueExtractionJob, tags: [jobs], requestBody: {$ref: '#/components/requestBodies/Provider'}, responses: {'200': {$ref: '#/components/responses/AgentJob'}, '202': {$ref: '#/components/responses/AgentJob'}, '400': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}, '502': {$ref: '#/components/responses/Error'}}}
  /api/modules:
    get: {operationId: listModules, tags: [modules], responses: {'200': {$ref: '#/components/responses/ModuleList'}}}
    post: {operationId: registerModule, tags: [modules], requestBody: {$ref: '#/components/requestBodies/RegisterModule'}, responses: {'201': {$ref: '#/components/responses/Module'}, '400': {$ref: '#/components/responses/Error'}, '422': {$ref: '#/components/responses/Error'}}}
  /api/modules/{moduleId}:
    parameters: [{$ref: '#/components/parameters/ModuleId'}]
    get: {operationId: getModule, tags: [modules], responses: {'200': {$ref: '#/components/responses/Module'}, '404': {$ref: '#/components/responses/Error'}}}
  /api/modules/{moduleId}/compare:
    parameters: [{$ref: '#/components/parameters/ModuleId'}]
    post: {operationId: compareModule, tags: [modules], requestBody: {$ref: '#/components/requestBodies/CompareModule'}, responses: {'200': {$ref: '#/components/responses/RegistryComparison'}, '400': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}}}
  /api/spec-enrichments:
    post: {operationId: createSpecEnrichment, tags: [spec-enrichments], requestBody: {$ref: '#/components/requestBodies/CreateSpecEnrichment'}, responses: {'201': {$ref: '#/components/responses/SpecEnrichment'}, '400': {$ref: '#/components/responses/Error'}}}
  /api/spec-enrichments/{enrichmentId}:
    parameters: [{$ref: '#/components/parameters/EnrichmentId'}]
    get: {operationId: getSpecEnrichment, tags: [spec-enrichments], responses: {'200': {$ref: '#/components/responses/SpecEnrichment'}, '404': {$ref: '#/components/responses/Error'}}}
  /api/spec-enrichments/{enrichmentId}/jobs:
    parameters: [{$ref: '#/components/parameters/EnrichmentId'}]
    post: {operationId: queueSpecEnrichmentJob, tags: [jobs], requestBody: {$ref: '#/components/requestBodies/Provider'}, responses: {'200': {$ref: '#/components/responses/AgentJob'}, '202': {$ref: '#/components/responses/AgentJob'}, '400': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}, '502': {$ref: '#/components/responses/Error'}}}
  /api/compositions:
    post: {operationId: createComposition, tags: [compositions], requestBody: {$ref: '#/components/requestBodies/CreateComposition'}, responses: {'201': {$ref: '#/components/responses/Composition'}, '400': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}}}
  /api/compositions/{compositionId}:
    parameters: [{$ref: '#/components/parameters/CompositionId'}]
    get: {operationId: getComposition, tags: [compositions], responses: {'200': {$ref: '#/components/responses/Composition'}, '404': {$ref: '#/components/responses/Error'}}}
  /api/compositions/{compositionId}/layout:
    parameters: [{$ref: '#/components/parameters/CompositionId'}]
    patch: {operationId: updateCompositionLayout, tags: [compositions], requestBody: {$ref: '#/components/requestBodies/PatchCompositionLayout'}, responses: {'200': {$ref: '#/components/responses/Composition'}, '400': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}}}
  /api/compositions/{compositionId}/clarification-jobs:
    parameters: [{$ref: '#/components/parameters/CompositionId'}]
    post: {operationId: queueCompositionClarificationJob, tags: [jobs], requestBody: {$ref: '#/components/requestBodies/Provider'}, responses: {'200': {$ref: '#/components/responses/AgentJob'}, '202': {$ref: '#/components/responses/AgentJob'}, '400': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}, '502': {$ref: '#/components/responses/Error'}}}
  /api/compositions/{compositionId}/answers:
    parameters: [{$ref: '#/components/parameters/CompositionId'}]
    post: {operationId: saveCompositionAnswers, tags: [compositions], requestBody: {$ref: '#/components/requestBodies/CompositionAnswers'}, responses: {'200': {$ref: '#/components/responses/Composition'}, '400': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}}}
  /api/compositions/{compositionId}/compile-jobs:
    parameters: [{$ref: '#/components/parameters/CompositionId'}]
    post: {operationId: queueCompositionCompileJob, tags: [jobs], requestBody: {$ref: '#/components/requestBodies/Provider'}, responses: {'200': {$ref: '#/components/responses/AgentJob'}, '202': {$ref: '#/components/responses/AgentJob'}, '400': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}, '502': {$ref: '#/components/responses/Error'}}}
...
  /api/workbench/validate-edge:
    post:
      operationId: validateWorkbenchEdge
      tags: [workbench]
      requestBody: {$ref: '#/components/requestBodies/ValidateEdge'}
      responses: {'200': {$ref: '#/components/responses/EdgeValidation'}, '400': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}, '422': {$ref: '#/components/responses/Error'}}
  /api/blueprints:
    post: {operationId: createBlueprint, tags: [blueprints], requestBody: {$ref: '#/components/requestBodies/Blueprint'}, responses: {'201': {$ref: '#/components/responses/Blueprint'}, '400': {$ref: '#/components/responses/Error'}, '422': {$ref: '#/components/responses/Error'}}}
    get: {operationId: listBlueprints, tags: [blueprints], responses: {'200': {$ref: '#/components/responses/BlueprintList'}}}
  /api/blueprints/{blueprintId}:
    parameters: [{$ref: '#/components/parameters/BlueprintId'}]
    get: {operationId: getBlueprint, tags: [blueprints], responses: {'200': {$ref: '#/components/responses/Blueprint'}, '404': {$ref: '#/components/responses/Error'}}}
    patch: {operationId: updateBlueprint, tags: [blueprints], requestBody: {$ref: '#/components/requestBodies/Blueprint'}, responses: {'200': {$ref: '#/components/responses/Blueprint'}, '400': {$ref: '#/components/responses/Error'}, '422': {$ref: '#/components/responses/Error'}}}
  /api/blueprints/{blueprintId}/validate:
    parameters: [{$ref: '#/components/parameters/BlueprintId'}]
    post: {operationId: validateBlueprint, tags: [blueprints], responses: {'200': {$ref: '#/components/responses/Blueprint'}, '404': {$ref: '#/components/responses/Error'}, '422': {$ref: '#/components/responses/Error'}}}
  /api/blueprints/{blueprintId}/wiring-jobs:
    parameters: [{$ref: '#/components/parameters/BlueprintId'}]
    post: {operationId: queueWiringJob, tags: [jobs], requestBody: {$ref: '#/components/requestBodies/Provider'}, responses: {'200': {$ref: '#/components/responses/AgentJob'}, '202': {$ref: '#/components/responses/AgentJob'}, '400': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}, '502': {$ref: '#/components/responses/Error'}}}
...
  /api/agent-jobs/{jobId}:
    parameters: [{$ref: '#/components/parameters/JobId'}]
    get: {operationId: getAgentJob, tags: [jobs], responses: {'200': {$ref: '#/components/responses/AgentJob'}, '404': {$ref: '#/components/responses/Error'}}}
  /api/agent-jobs/{jobId}/open:
    parameters: [{$ref: '#/components/parameters/JobId'}]
    post: {operationId: openAgentJob, tags: [jobs], responses: {'200': {$ref: '#/components/responses/JobOpen'}, '404': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}, '502': {$ref: '#/components/responses/Error'}}}
  /api/agent-jobs/{jobId}/cancel:
    parameters: [{$ref: '#/components/parameters/JobId'}]
    post: {operationId: cancelAgentJob, tags: [jobs], responses: {'200': {$ref: '#/components/responses/AgentJob'}, '404': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}, '502': {$ref: '#/components/responses/Error'}}}
components:
  parameters:
    SessionId: {name: sessionId, in: path, required: true, schema: {type: string}}
    CandidateId: {name: candidateId, in: path, required: true, schema: {type: string}}
    PlanId: {name: planId, in: path, required: true, schema: {type: string}}
    ModuleId: {name: moduleId, in: path, required: true, schema: {type: string}}
...
    EnrichmentId: {name: enrichmentId, in: path, required: true, schema: {type: string}}
...
    RelativePath: {name: path, in: query, required: true, schema: {type: string, not: {pattern: '(^/|\\.\\.|\\\\)'}}}
    SessionFilter: {name: sessionId, in: query, required: false, schema: {type: string}}
    RepositoryFilter: {name: repositoryId, in: query, required: false, schema: {type: string}}
    StatusFilter: {name: status, in: query, required: false, schema: {type: string}}
    RiskFilter: {name: extractionRisk, in: query, required: false, schema: {type: string}}
    ConfidenceFilter: {name: confidence, in: query, required: false, schema: {type: string}}
    CapabilityFilter: {name: capability, in: query, required: false, schema: {type: string}}
  responses:
...
    Repository: {description: Repository response, content: {application/json: {schema: {$ref: '#/components/schemas/Repository'}}}}
    RepositoryList: {description: Repository list response, content: {application/json: {schema: {$ref: '#/components/schemas/RepositoryList'}}}}
    Session: {description: Session response, content: {application/json: {schema: {$ref: '#/components/schemas/Session'}}}}
    SessionList: {description: Session list response, content: {application/json: {schema: {$ref: '#/components/schemas/SessionList'}}}}
    SourceFile: {description: Source file response, content: {application/json: {schema: {$ref: '#/components/schemas/SourceFile'}}}}
    Candidate: {description: Candidate response, content: {application/json: {schema: {$ref: '#/components/schemas/Candidate'}}}}
    CandidateList: {description: Candidate list response, content: {application/json: {schema: {$ref: '#/components/schemas/CandidateList'}}}}
    ExtractionPlan: {description: Extraction plan response, content: {application/json: {schema: {$ref: '#/components/schemas/ExtractionPlan'}}}}
    Module: {description: Module response, content: {application/json: {schema: {$ref: '#/components/schemas/Module'}}}}
    ModuleList: {description: Module list response, content: {application/json: {schema: {$ref: '#/components/schemas/ModuleList'}}}}
...
    AgentJob: {description: Agent job response, content: {application/json: {schema: {$ref: '#/components/schemas/AgentJob'}}}}
...
    Error: {description: Error response, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}}
  requestBodies:
    CreateRepository: {required: true, content: {application/json: {schema: {type: object, additionalProperties: false, required: [sourceType, sourceUri], properties: {name: {type: string}, sourceType: {enum: [local_path, git_url]}, sourceUri: {type: string}, defaultBranch: {type: string}, rescan: {type: boolean}}}}}}
    CreateSession: {required: true, content: {application/json: {schema: {type: object, additionalProperties: false, required: [repositoryId], properties: {repositoryId: {type: string}}}}}}
    SessionIntent: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/SessionIntent'}}}}
    PatchCandidate: {required: true, content: {application/json: {schema: {type: object, additionalProperties: false, properties: {proposedName: {type: string}, description: {type: string}, moduleKind: {type: string}}}}}}
    Reason: {required: true, content: {application/json: {schema: {type: object, additionalProperties: false, required: [reason], properties: {reason: {type: string, minLength: 3}, duplicateModuleId: {type: string}}}}}}
    OptionalReason: {required: true, content: {application/json: {schema: {type: object, additionalProperties: false, properties: {reason: {type: string}, duplicateModuleId: {type: string}}}}}}
    Provider: {required: true, content: {application/json: {schema: {type: object, additionalProperties: false, properties: {provider: {type: string, enum: [claude_code_tmux, fake]}}}}}}
    CompareModule: {required: true, content: {application/json: {schema: {type: object, additionalProperties: false, properties: {compareToModuleId: {type: string}}}}}}
    CreateExtractionPlan: {required: true, content: {application/json: {schema: {type: object, additionalProperties: false, required: [sessionId, approvedCandidateIds], properties: {sessionId: {type: string}, approvedCandidateIds: {type: array, items: {type: string}}, rejectedCandidateIds: {type: array, items: {type: string}}}}}}}
...